# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "WhatsApp HTTP API",
    "description": "WhatsApp HTTP API that you can configure in a click",
    "version": "1.0",
    "contact": {}
  },
  "tags": [
    {
      "name": "device",
      "description": "Device information"
    },
    {
      "name": "chatting",
      "description": "Chat methods"
    }
  ],
  "servers": [],
  "components": {
    "securitySchemes": {
      "api_key": {
        "type": "apiKey",
        "in": "header",
        "name": "X-VENOM-TOKEN",
        "description": "Your secret key"
      }
    },
    "schemas": {
      "MessageContactVcard": {
        "type": "object",
        "properties": {
          "chatId": {
            "type": "string",
            "example": "791231234567@c.us"
          },
          "contactsId": {
            "type": "string"
          },
          "name": {
            "type": "string"
          }
        },
        "required": [
          "chatId",
          "contactsId",
          "name"
        ]
      },
      "MessageText": {
        "type": "object",
        "properties": {
          "chatId": {
            "type": "string",
            "example": "791231234567@c.us"
          },
          "text": {
            "type": "string"
          }
        },
        "required": [
          "chatId",
          "text"
        ]
      },
      "MessageLocation": {
        "type": "object",
        "properties": {
          "chatId": {
            "type": "string",
            "example": "791231234567@c.us"
          },
          "latitude": {
            "type": "string"
          },
          "longitude": {
            "type": "string"
          },
          "title": {
            "type": "string"
          }
        },
        "required": [
          "chatId",
          "latitude",
          "longitude",
          "title"
        ]
      },
      "MessageLinkPreview": {
        "type": "object",
        "properties": {
          "chatId": {
            "type": "string",
            "example": "791231234567@c.us"
          },
          "url": {
            "type": "string"
          },
          "title": {
            "type": "string"
          }
        },
        "required": [
          "chatId",
          "url",
          "title"
        ]
      },
      "MessageImage": {
        "type": "object",
        "properties": {
          "chatId": {
            "type": "string",
            "example": "791231234567@c.us"
          },
          "path": {
            "type": "string"
          },
          "filename": {
            "type": "string"
          },
          "caption": {
            "type": "string"
          }
        },
        "required": [
          "chatId",
          "path",
          "filename",
          "caption"
        ]
      },
      "MessageFile": {
        "type": "object",
        "properties": {
          "chatId": {
            "type": "string",
            "example": "791231234567@c.us"
          },
          "path": {
            "type": "string"
          },
          "filename": {
            "type": "string"
          },
          "caption": {
            "type": "string"
          }
        },
        "required": [
          "chatId",
          "path",
          "filename",
          "caption"
        ]
      },
      "MessageReply": {
        "type": "object",
        "properties": {
          "chatId": {
            "type": "string",
            "example": "791231234567@c.us"
          },
          "reply_to": {
            "type": "string",
            "example": "message.id"
          },
          "text": {
            "type": "string"
          }
        },
        "required": [
          "chatId",
          "reply_to",
          "text"
        ]
      },
      "Chat": {
        "type": "object",
        "properties": {
          "chatId": {
            "type": "string",
            "example": "791231234567@c.us"
          }
        },
        "required": [
          "chatId"
        ]
      }
    }
  },
  "paths": {
    "/api/sendContactVcard": {
      "post": {
        "operationId": "ChattingController_sendContactVcard",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MessageContactVcard"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "tags": [
          "chatting"
        ]
      }
    },
    "/api/sendText": {
      "post": {
        "operationId": "ChattingController_sendText",
        "summary": "Send a text message",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MessageText"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "tags": [
          "chatting"
        ]
      }
    },
    "/api/sendLocation": {
      "post": {
        "operationId": "ChattingController_sendLocation",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MessageLocation"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "tags": [
          "chatting"
        ]
      }
    },
    "/api/sendLinkPreview": {
      "post": {
        "operationId": "ChattingController_sendLinkPreview",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MessageLinkPreview"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "tags": [
          "chatting"
        ]
      }
    },
    "/api/sendImage": {
      "post": {
        "operationId": "ChattingController_sendImage",
        "summary": "NOT IMPLEMENTED YET",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MessageImage"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "tags": [
          "chatting"
        ]
      }
    },
    "/api/sendFile": {
      "post": {
        "operationId": "ChattingController_sendFile",
        "summary": "NOT IMPLEMENTED YET",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MessageFile"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "tags": [
          "chatting"
        ]
      }
    },
    "/api/reply": {
      "post": {
        "operationId": "ChattingController_reply",
        "summary": "Reply to a text message",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MessageReply"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "tags": [
          "chatting"
        ]
      }
    },
    "/api/sendSeen": {
      "post": {
        "operationId": "ChattingController_sendSeen",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Chat"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "chatting"
        ]
      }
    },
    "/api/startTyping": {
      "post": {
        "operationId": "ChattingController_startTyping",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Chat"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "boolean"
                }
              }
            }
          }
        },
        "tags": [
          "chatting"
        ]
      }
    },
    "/api/stopTyping": {
      "post": {
        "operationId": "ChattingController_stopTyping",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Chat"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "boolean"
                }
              }
            }
          }
        },
        "tags": [
          "chatting"
        ]
      }
    },
    "/api/killServiceWorker": {
      "post": {
        "operationId": "DeviceController_killServiceWorker",
        "parameters": [],
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "boolean"
                }
              }
            }
          }
        },
        "tags": [
          "device"
        ]
      }
    },
    "/api/restartService": {
      "post": {
        "operationId": "DeviceController_restartService",
        "parameters": [],
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "boolean"
                }
              }
            }
          }
        },
        "tags": [
          "device"
        ]
      }
    },
    "/api/getHostDevice": {
      "get": {
        "operationId": "DeviceController_getHostDevice",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "tags": [
          "device"
        ]
      }
    },
    "/api/getConnectionState": {
      "get": {
        "operationId": "DeviceController_getConnectionState",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "device"
        ]
      }
    },
    "/api/getBatteryLevel": {
      "get": {
        "operationId": "DeviceController_getBatteryLevel",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "number"
                }
              }
            }
          }
        },
        "tags": [
          "device"
        ]
      }
    },
    "/api/isConnected": {
      "get": {
        "operationId": "DeviceController_isConnected",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "boolean"
                }
              }
            }
          }
        },
        "tags": [
          "device"
        ]
      }
    },
    "/api/getWAVersion": {
      "get": {
        "operationId": "DeviceController_getWAVersion",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "tags": [
          "device"
        ]
      }
    },
    "/api/screenshot": {
      "get": {
        "operationId": "ScreenshotController_screenshot",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "screenshot"
        ]
      }
    }
  }
}